get rid of last awkward condition, add tests
